PlugPopup : Add new class for editing plugs in a popup window
This consolidates existing functionality from _SceneViewInspector and SpreadsheetUI._EditWindow into a single class we can use elsewhere.
