+++ python/GafferUI/__init__.py
@@ -185,6 +185,7 @@ from .SplineWidget import SplineWidget
 from .Bookmarks import Bookmarks
 from . import WidgetAlgo
 from .CodeWidget import CodeWidget
+from .PlugPopup import PlugPopup
 
 # then all the PathPreviewWidgets. note that the order
 # of import controls the order of display.
